q2.Ferris Wheel
There are n children who want to go to a Ferris wheel, and your task is to find a gondola for each child.

Each gondola may have one or two children in it, and in addition, the total weight in a gondola may not exceed x. You know the weight of every child.

What is the minimum number of gondolas needed for the children? Input

The first input line contains two integers n and x: the number of children and the maximum allowed weight. e.g., Input:
4 10 ; n=4, x=10
7 2 3 9 ; p[n]=7,2,3,9.....